Update arpSpoof.py
Hey stylish-bear!

I'm referencing the code found on the publisher's link to Github:
https://github.com/The-Ethical-Hacking-Book/Code-by-chapter/blob/main/Chapter-02/arpSpoof.py

In the book, it says "Implement the arp_spoof() function. This function should be very similar to arp_restore(), which restores the ARP tables to their original state. You can use arp_restore() as a guide."

Based on this, I think the operator "who-is" should be changed to "is-at", as is in the arp_restore() function which switches the ARP tables back to normal to avoid detection after ending the arp spoof attack. 
---

def arp_spoof(dest_ip, dest_mac, source_ip):
    packet = ARP(op="who-has", psrc = source_ip,
            hwdst = dest_mac, pdst = dest_ip)
    send(packet, verbose = False)

---
When arp_spoof() is called, it should send a packet, updating the ARP table at the destination. The packet should tell (using op="as-is") the destination that the MAC address of source_ip, is to be updated to the attacking device's MAC address (using hwsrc=get_if_hwaddr(conf.iface)). 

So now, any future traffic addressed from the destination to source_ip is incorrectly delivered to the attacking device, because the ARP table now shows attacking device's MAC address as the final destination.

I think the suggested operator change is verified to work when when the code later runs--

try:
        print("Sending spoofed ARP packets")
        while True:
            arp_spoof(victim_ip, victim_mac, router_ip)
            arp_spoof(router_ip, router_mac, victim_ip)

--it first sends a message to the victim that the router's IP is now associated with the attacker's MAC. Next, it tells the router that the victim's IP is now associated with the attacker's MAC. By switching the MAC at both destinations, the attacking device is now intercepting.

Let me know what you think. I'm trying to check my understanding by looking at other people's work from this book.